+++ SandboxiePlus/SandMan/Windows/OptionsGeneral.cpp
@@ -892,12 +892,12 @@ void COptionsWindow::OnBrowsePath()
 	if (Value.isEmpty())
 		return;
 
-	QString Name = QInputDialog::getText(this, "Sandboxie-Plus", tr("Please enter a menu title"), QLineEdit::Normal);
+	QString Name = QInputDialog::getText(this, "Sandboxie-Plus", tr("Please enter a menu title"), QLineEdit::Normal, Split2(Value, "\\", true).second);
 	if (Name.isEmpty())
 		return;
 
 	CSandBoxPlus* pBoxEx = qobject_cast<CSandBoxPlus*>(m_pBox.data());
-	AddRunItem(Name, "", "\"" + (pBoxEx ? Value.replace(pBoxEx->GetFileRoot(), "%BoxRoot%", Qt::CaseInsensitive) : Value) + "\"");
+	AddRunItem(Name, "", "\"" + (pBoxEx ? pBoxEx->MakeBoxCommand(Value) : Value) + "\"");
 	m_GeneralChanged = true;
 	OnOptChanged();
 }